+++ packages/inspector/src/types/arguments.ts
@@ -10,6 +10,6 @@ const inspectArguments: InspectType<IArguments> = (arguments_: IArguments, optio
     options.truncate -= 13;
 
     return `Arguments [ ${inspectList(arguments_, arguments_, options, inspect)} ]`;
-}
+};
 
 export default inspectArguments;