+++ packages/pail/__bench__/object.bench.ts
@@ -8,10 +8,10 @@ import { createConsola as createBrowserConsola } from "consola/browser";
 import * as winston from "winston";
 import bunyan from "bunyan";
 
-import { createPail as createServerPail } from "../src/index.server";
-import { createPail as createBrowserPail } from "../src/index.browser";
-import { JsonReporter as ServerJsonReporter } from "../src/reporter/json/json.server";
-import { JsonReporter as BrowserJsonReporter } from "../src/reporter/json/json.browser";
+import { createPail as createServerPail } from "@visulima/pail/server";
+import { createPail as createBrowserPail } from "@visulima/pail/browser";
+import { JsonReporter as ServerJsonReporter } from "@visulima/pail/server/reporter";
+import { JsonReporter as BrowserJsonReporter } from "@visulima/pail/browser/reporter";
 import { JsonBrowserConsolaReporter, JsonServerConsolaReporter } from "./utils";
 
 const wsDevNull = fs.createWriteStream("/dev/null");