+++ packages/pail/__tests__/unit/reporter/simple/simple.server.test.ts
@@ -1,11 +1,11 @@
 import { stderr, stdout } from "node:process";
 
 import { blueBright, bold, grey, red } from "@visulima/colorize";
-import { stringify } from "safe-stable-stringify";
 import { describe, expect, it, vi } from "vitest";
 
 import { dateFormatter } from "../../../../src/reporter/pretty/abstract-pretty-reporter";
 import SimpleReporter from "../../../../src/reporter/simple/simple.server";
+import type { Meta, ReadonlyMeta } from "../../../../src/types";
 
 vi.mock("terminal-size", () => {
     return {
@@ -30,17 +30,18 @@ describe("simpleReporter", () => {
             date,
             groups: ["Group1"],
             label: "Label",
-            message: "This is a sample message",
+            message: "This is a sample message1",
             type: {
                 level: "informational",
                 name: "info",
             },
         };
-        const stdoutSpy = vi.spyOn(stdout, "write");
-        simpleReporter.log(meta);
+        const stdoutSpy = vi.spyOn(stdout, "write").mockImplementation(() => true);
+
+        simpleReporter.log(meta as ReadonlyMeta<string>);
 
         expect(stdoutSpy).toHaveBeenCalledWith(
-            `    ${grey("[Group1]") + " " + grey(dateFormatter(date))} ${bold(blueBright("INFO")) + bold(blueBright("LABEL"))}         This is a sample message\n`,
+            `    ${grey("[Group1]") + " " + grey(dateFormatter(date))} ${bold(blueBright("INFO")) + bold(blueBright("LABEL"))}         This is a sample message1\n`,
         );
 
         stdoutSpy.mockRestore();
@@ -62,8 +63,9 @@ describe("simpleReporter", () => {
                 name: "info",
             },
         };
-        const stdoutSpy = vi.spyOn(stdout, "write");
-        simpleReporter.log(meta);
+        const stdoutSpy = vi.spyOn(stdout, "write").mockImplementation(() => true);
+
+        simpleReporter.log(meta as ReadonlyMeta<string>);
 
         expect(stdoutSpy).toHaveBeenCalledWith(
             `    ${grey("[Group1]") + " " + grey(dateFormatter(date))} ${bold(blueBright("INFO")) + bold(blueBright("LABEL"))}     ${grey("[Scope1 > Scope2]")}     This is a sample message\n`,
@@ -89,7 +91,7 @@ describe("simpleReporter", () => {
         };
         const stderrSpy = vi.spyOn(stderr, "write").mockImplementation(() => true);
 
-        simpleReporter.log(meta);
+        simpleReporter.log(meta as ReadonlyMeta<string>);
 
         expect(stderrSpy).toHaveBeenCalledWith(
             `    ${grey("[Group1]") + " " + grey(dateFormatter(date))} ${bold(red("ERROR")) + bold(red("LABEL"))}         This is an error message\n`,
@@ -118,7 +120,7 @@ describe("simpleReporter", () => {
             },
         };
 
-        simpleReporter.log(meta);
+        simpleReporter.log(meta as ReadonlyMeta<string>);
 
         // eslint-disable-next-line @typescript-eslint/unbound-method
         expect(newStdout.write).toHaveBeenCalledWith(
@@ -146,7 +148,7 @@ describe("simpleReporter", () => {
             },
         };
 
-        simpleReporter.log(meta);
+        simpleReporter.log(meta as ReadonlyMeta<string>);
 
         // eslint-disable-next-line @typescript-eslint/unbound-method
         expect(newStderr.write).toHaveBeenCalledWith(
@@ -158,7 +160,6 @@ describe("simpleReporter", () => {
         expect.assertions(1);
 
         const simpleReporter = new SimpleReporter();
-        simpleReporter.setStringify(stringify);
 
         const meta = {
             badge: undefined,
@@ -173,7 +174,7 @@ describe("simpleReporter", () => {
         };
         const stdoutSpy = vi.spyOn(stdout, "write").mockImplementation(() => true);
 
-        expect(() => simpleReporter.log(meta)).not.toThrow();
+        expect(() => simpleReporter.log(meta as unknown as Meta<string>)).not.toThrow();
 
         stdoutSpy.mockRestore();
     });
@@ -196,7 +197,7 @@ describe("simpleReporter", () => {
         };
 
         // @ts-expect-error - just for testing
-        const formattedMessage = simpleReporter._formatMessage(meta);
+        const formattedMessage = simpleReporter._formatMessage(meta as ReadonlyMeta<string>);
 
         expect(formattedMessage).toContain("This is a sample message");
     });
@@ -216,8 +217,35 @@ describe("simpleReporter", () => {
                 name: "info",
             },
         };
-        const formattedMessage = simpleReporter._formatMessage(meta);
+        // @ts-expect-error - just for testing
+        const formattedMessage = simpleReporter._formatMessage(meta as ReadonlyMeta<string>);
 
         expect(formattedMessage.split("\n").length).toBeGreaterThan(1);
+    });
+
+    it("should not strip spaces from messages", () => {
+        expect.assertions(1);
+
+        const simpleReporter = new SimpleReporter();
+        const meta = {
+            badge: "INFO",
+            date,
+            groups: ["Group1"],
+            label: "Label",
+            message: "  a  This is a sample message1",
+            type: {
+                level: "informational",
+                name: "info",
+            },
+        };
+        const stdoutSpy = vi.spyOn(stdout, "write").mockImplementation(() => true);
+
+        simpleReporter.log(meta as ReadonlyMeta<string>);
+
+        expect(stdoutSpy).toHaveBeenCalledWith(
+            `    ${grey("[Group1]") + " " + grey(dateFormatter(date))} ${bold(blueBright("INFO")) + bold(blueBright("LABEL"))}           a  This is a sample message1\n`,
+        );
+
+        stdoutSpy.mockRestore();
     });
 });