+++ packages/pail/src/reporter/file/json-file-reporter.ts
@@ -1,6 +1,7 @@
 import type { stringify } from "safe-stable-stringify";
 
-import type { Meta } from "../../types";
+import { EMPTY_SYMBOL } from "../../constants";
+import type { ReadonlyMeta } from "../../types";
 import type { Options as FileReporterOptions } from "./abstract-file-reporter";
 import { AbstractFileReporter } from "./abstract-file-reporter";
 
@@ -22,10 +23,10 @@ class JsonFileReporter<L extends string = string> extends AbstractFileReporter<L
         this.#stringify = function_;
     }
 
-    protected _formatMessage(meta: Meta<L>): string {
+    protected _formatMessage(meta: ReadonlyMeta<L>): string {
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment,@typescript-eslint/prefer-ts-expect-error
         // @ts-ignore - @TODO: check rollup-plugin-dts
-        const { file, type, ...rest } = meta;
+        const { file, message, type, ...rest } = meta;
 
         if (rest.label) {
             rest.label = rest.label.trim();
@@ -35,7 +36,13 @@ class JsonFileReporter<L extends string = string> extends AbstractFileReporter<L
         // @ts-nocheck - @TODO: check rollup-plugin-dts
         if (file) {
             // This is a hack to make the file property a string
-            (rest as unknown as Omit<Meta<L>, "file"> & { file: string }).file = file.name + ":" + file.line + (file.column ? ":" + file.column : "");
+            (rest as unknown as Omit<ReadonlyMeta<L>, "file"> & { file: string }).file = file.name + ":" + file.line + (file.column ? ":" + file.column : "");
+        }
+
+        if (message === EMPTY_SYMBOL) {
+            (rest as unknown as Omit<ReadonlyMeta<L>, "message"> & { message: string | undefined }).message = undefined;
+        } else {
+            (rest as unknown as Omit<ReadonlyMeta<L>, "message"> & { message: ReadonlyMeta<L>["message"] }).message = message;
         }
 
         return (this.#stringify as typeof stringify)(rest) as string;